Moved role creation to the success popup
